Service class FileV8Reader implements only reading functions without deflating
